+++ src/main/vuejs/components/not-found.vue
@@ -1,3 +1,8 @@
 <template>
-  <h1>Page not found :/</h1>
+  <div class="jumbotron jumbotron-fluid">
+    <div class="container">
+      <h1 class="display-3"><i class="fa fa-meh-o"></i> Error 404</h1>
+      <p class="lead">The page You are looking for is not here</p>
+    </div>
+  </div>
 </template>