+++ src/main/vuejs/components/user/index.vue
@@ -1,7 +1,8 @@
 <template>
-  <div>
-    <h2>User page:</h2>
-    <i class="bug icon"></i>
-    <router-view></router-view>
+  <div class="jumbotron jumbotron-fluid">
+    <div class="container">
+      <h1 class="display-3"><i class="fa fa-users"></i>User page</h1>
+      <router-view></router-view>
+    </div>
   </div>
 </template>